requirements.txt
#requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-community
sqlite-utils
python-dotenv
pydantic


